requirements.txt
aiofiles==22.1.0
aiosqlite==0.18.0
altair==4.2.2
anyio==3.6.2
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==22.2.0
Babel==2.12.1
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.0
bleach==6.0.0
blinker==1.4
branca==0.6.0
Brlapi==0.8.3
cachetools==5.3.0
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.4
comm==0.1.3
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
debugpy==1.6.6
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
dell-recovery==0.0.0
DellLinuxAssistant==0.0.0
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
duplicity==0.8.21
entrypoints==0.4
executing==1.2.0
fasteners==0.14.1
fastjsonschema==2.16.3
filelock==3.10.7
folium==0.14.0
fqdn==1.5.1
future==0.18.2
gitdb==4.0.10
GitPython==3.1.31
haversine==2.8.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-console==6.6.3
jupyter-contrib-core==0.4.2
jupyter-contrib-nbextensions==0.7.0
jupyter-events==0.6.3
jupyter-highlight-selected-word==0.2.0
jupyter-nbextensions-configurator==0.6.1
jupyter-ydoc==0.2.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_fileid==0.9.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
jupyterlab_server==2.22.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.9.2
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==2.2.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
monotonic==1.6
more-itertools==8.10.0
nbclassic==0.5.4
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.8.0
nest-asyncio==1.5.6
netifaces==0.11.0
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.24.2
oauthlib==3.2.0
olefile==0.46
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==3.2.0
plotly==5.14.1
progressbar==2.5
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.12.4
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydeck==0.8.0
Pygments==2.14.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
Pympler==1.0.1
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
PyQtWebEngine==5.15.5
pyRFC3339==1.1
pyrsistent==0.19.3
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-json-logger==2.0.7
python-magic==0.4.24
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.0.2
qtconsole==5.4.1
QtPy==2.3.1
reportlab==3.6.8
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.3
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sos==4.4
soupsieve==2.4
stack-data==0.6.2
streamlit==1.21.0
streamlit-folium==0.11.1
systemd-python==234
tenacity==8.2.2
terminado==0.17.1
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.2.0
urllib3==1.26.5
usb-creator==0.3.7
validators==0.20.0
virtualenv==20.21.0
wadllib==1.3.6
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
widgetsnbextension==4.0.7
xdg==5
xkit==0.0.0
y-py==0.5.9
ypy-websocket==0.8.2
zipp==1.0.0
